+++ src/controllers/expenses.controller.js
@@ -1,10 +1,10 @@
 
 const getMonth = require('../utils/getMonth')
-const expenseModel = require('../../models/expense.model');
-const monthModel = require('../../models/month.model');
+const expenseModel = require('../models/expense.model');
+const monthModel = require('../models/month.model');
 const queryExpensesOfAType = require('../utils/queryExpensesOfAType');
-const incomeModel = require('../../models/income.model');
-const setUpModel = require('../../models/setup.model');
+const incomeModel = require('../models/income.model');
+const setUpModel = require('../models/setup.model');
 const {ObjectId} = require("mongodb");
 
 /**
